Add getPhases function to a helper file
